refactor: rename columns to cabinColumns
Des:
change direction and rename
